+++ src/activation.h
@@ -5,9 +5,9 @@
 # include <stdio.h>
 # endif
 
-float activation_relu_cpu(float x);
-__device__ float activation_relu(float x);
-__global__ void forward_relu(float *input, float *output, const int size);
+double activation_relu_cpu(double x);
+__device__ double activation_relu(double x);
+__global__ void forward_relu(double *input, double *output, const int size);
 
-float* cpu_relu(float *input, const int size, bool inplace = true);
-float* relu(dim3 grid, dim3 block, float *input, const int size, bool inplace = true);
+double* cpu_relu(double *input, const int size, bool inplace = true);
+double* relu(dim3 grid, dim3 block, double *input, const int size, bool inplace = true);
